move handleSetupKVClient to its own file
