+++ src/hooks.server.ts
@@ -1,20 +1,9 @@
 import type { Handle } from "@sveltejs/kit";
-import { CONTENTFUL_DELIVERY_API_TOKEN, CONTENTFUL_SPACE_ID, KV_URL } from "$env/static/private";
+import { CONTENTFUL_DELIVERY_API_TOKEN, CONTENTFUL_SPACE_ID } from "$env/static/private";
 import { sequence } from "@sveltejs/kit/hooks";
-import { getClient as getKVClient } from "$lib/services/server/kv";
 import getContentfulClient from "$lib/services/server/contentful";
 
-const handleSetupKVClient = (async ({ event, resolve }) => {
-  // we intentionally don't await this promise here, so that other things can happen while redis is
-  // initializing. Anything that needs redis can await the promise by calling
-  // event.locals.getKVClient and awaiting the result. If nothing awaits the promise
-  // then we never wait on redis connecting first.
-  const kvClientPromise = KV_URL
-    ? getKVClient({ url: KV_URL })
-    : Promise.reject(new Error("could not get KV client; no KV_URL was specified"));
-  event.locals.getKVClient = () => kvClientPromise;
-  return resolve(event);
-}) satisfies Handle;
+import handleSetupKVClient from "./hooks/server/handleSetupKVClient";
 
 const handleSetupContentfulClient = (async ({ event, resolve }) => {
   const { fetch } = event;
